doc + update notify + migrate sqlite to postgersql

+++ src/storage/schema.sql
@@ -1,18 +1,18 @@
 CREATE TABLE IF NOT EXISTS candles (
     symbol TEXT NOT NULL,
     timeframe TEXT NOT NULL,
-    open_time INTEGER NOT NULL,
+    open_time BIGINT NOT NULL,
     open REAL NOT NULL,
     high REAL NOT NULL,
     low REAL NOT NULL,
     close REAL NOT NULL,
     volume REAL NOT NULL,
-    close_time INTEGER NOT NULL,
+    close_time BIGINT NOT NULL,
     UNIQUE(symbol, timeframe, open_time)
 );
 
 CREATE TABLE IF NOT EXISTS signals (
-    ts INTEGER NOT NULL,
+    ts BIGINT NOT NULL,
     symbol TEXT NOT NULL,
     timeframe TEXT NOT NULL,
     signal TEXT NOT NULL,
@@ -23,11 +23,11 @@ CREATE TABLE IF NOT EXISTS signals (
 
 CREATE TABLE IF NOT EXISTS trades (
     trade_id TEXT PRIMARY KEY,
-    ts INTEGER NOT NULL,
+    ts BIGINT NOT NULL,
     side TEXT NOT NULL,
     symbol TEXT NOT NULL,
     timeframe TEXT NOT NULL,
-    source_open_time INTEGER NOT NULL,
+    source_open_time BIGINT NOT NULL,
     qty_btc REAL NOT NULL,
     price_market REAL NOT NULL,
     price_exec REAL NOT NULL,
@@ -41,7 +41,7 @@ CREATE TABLE IF NOT EXISTS trades (
 );
 
 CREATE TABLE IF NOT EXISTS equity (
-    ts INTEGER NOT NULL,
+    ts BIGINT NOT NULL,
     cash REAL NOT NULL,
     btc_qty REAL NOT NULL,
     btc_price REAL NOT NULL,
